+++ src/visor/visor.go
@@ -90,18 +90,6 @@ type Config struct {
 	//Secret key of blockchain authority (if master)
 	BlockchainSeckey cipher.SecKey
 
-	// How often new blocks are created by the master, in seconds
-	BlockCreationInterval uint64
-	// How often an unconfirmed txn is checked against the blockchain
-	UnconfirmedCheckInterval time.Duration
-	// How long we'll hold onto an unconfirmed txn
-	UnconfirmedMaxAge time.Duration
-	// How often to check the unconfirmed pool for transactions that become valid
-	UnconfirmedRefreshRate time.Duration
-	// How often to remove transactions that become permanently invalid from the unconfirmed pool
-	UnconfirmedRemoveInvalidRate time.Duration
-	// How often to rebroadcast unconfirmed transactions
-	UnconfirmedResendPeriod time.Duration
 	// Maximum size of a block, in bytes.
 	MaxBlockSize int
 
@@ -144,15 +132,7 @@ func NewVisorConfig() Config {
 		BlockchainPubkey: cipher.PubKey{},
 		BlockchainSeckey: cipher.SecKey{},
 
-		BlockCreationInterval: 10,
-		//BlockCreationForceInterval: 120, //create block if no block within this many seconds
-
-		UnconfirmedCheckInterval:     time.Hour * 2,
-		UnconfirmedMaxAge:            time.Hour * 48,
-		UnconfirmedRefreshRate:       time.Minute,
-		UnconfirmedRemoveInvalidRate: time.Minute,
-		UnconfirmedResendPeriod:      time.Minute,
-		MaxBlockSize:                 DefaultMaxBlockSize,
+		MaxBlockSize: DefaultMaxBlockSize,
 
 		GenesisAddress:    cipher.Address{},
 		GenesisSignature:  cipher.Sig{},
@@ -174,9 +154,12 @@ func (c Config) Verify() error {
 	return nil
 }
 
-// historyer is the interface that provides methods for accessing history data that are parsed from blockchain.
-type historyer interface {
-	GetUxOut(tx *dbutil.Tx, uxid cipher.SHA256) (*historydb.UxOut, error)
+//go:generate go install
+//go:generate goautomock -template=testify Historyer
+
+// Historyer is the interface that provides methods for accessing history data that are parsed from blockchain.
+type Historyer interface {
+	GetUxOuts(tx *dbutil.Tx, uxids []cipher.SHA256) ([]*historydb.UxOut, error)
 	ParseBlock(tx *dbutil.Tx, b coin.Block) error
 	GetTransaction(tx *dbutil.Tx, hash cipher.SHA256) (*historydb.Transaction, error)
 	GetAddrUxOuts(tx *dbutil.Tx, address cipher.Address) ([]*historydb.UxOut, error)
@@ -203,7 +186,7 @@ type Blockchainer interface {
 	ExecuteBlock(tx *dbutil.Tx, sb *coin.SignedBlock) error
 	VerifyBlockTxnConstraints(tx *dbutil.Tx, txn coin.Transaction) error
 	VerifySingleTxnHardConstraints(tx *dbutil.Tx, txn coin.Transaction) error
-	VerifySingleTxnAllConstraints(tx *dbutil.Tx, txn coin.Transaction, maxSize int) error
+	VerifySingleTxnSoftHardConstraints(tx *dbutil.Tx, txn coin.Transaction, maxSize int) error
 	TransactionFee(tx *dbutil.Tx, hours uint64) coin.FeeCalculator
 }
 
@@ -238,7 +221,7 @@ type Visor struct {
 	Wallets     *wallet.Service
 	StartedAt   time.Time
 
-	history historyer
+	history Historyer
 }
 
 // NewVisor creates a Visor for managing the blockchain database
@@ -492,7 +475,7 @@ func (vs *Visor) createBlock(tx *dbutil.Tx, when uint64) (coin.SignedBlock, erro
 	// Filter transactions that violate all constraints
 	var filteredTxns coin.Transactions
 	for _, txn := range txns {
-		if err := vs.Blockchain.VerifySingleTxnAllConstraints(tx, txn, vs.Config.MaxBlockSize); err != nil {
+		if err := vs.Blockchain.VerifySingleTxnSoftHardConstraints(tx, txn, vs.Config.MaxBlockSize); err != nil {
 			switch err.(type) {
 			case ErrTxnViolatesHardConstraint, ErrTxnViolatesSoftConstraint:
 				logger.Warningf("Transaction %s violates constraints: %v", txn.TxIDHex(), err)
@@ -830,10 +813,14 @@ func (vs *Visor) InjectTransaction(txn coin.Transaction) (bool, *ErrTxnViolatesS
 // The bool return value is whether or not the transaction was already in the pool.
 // If the transaction violates hard or soft constraints, it is rejected, and error will not be nil.
 func (vs *Visor) InjectTransactionStrict(txn coin.Transaction) (bool, error) {
+	if err := VerifySingleTxnUserConstraints(txn); err != nil {
+		return false, err
+	}
+
 	var known bool
 
 	if err := vs.DB.Update("InjectTransactionStrict", func(tx *dbutil.Tx) error {
-		err := vs.Blockchain.VerifySingleTxnAllConstraints(tx, txn, vs.Config.MaxBlockSize)
+		err := vs.Blockchain.VerifySingleTxnSoftHardConstraints(tx, txn, vs.Config.MaxBlockSize)
 		if err != nil {
 			return err
 		}
@@ -1457,17 +1444,21 @@ func (vs Visor) GetHeadBlockTime() (uint64, error) {
 
 // GetUxOutByID gets UxOut by hash id.
 func (vs Visor) GetUxOutByID(id cipher.SHA256) (*historydb.UxOut, error) {
-	var out *historydb.UxOut
+	var outs []*historydb.UxOut
 
 	if err := vs.DB.View("GetUxOutByID", func(tx *dbutil.Tx) error {
 		var err error
-		out, err = vs.history.GetUxOut(tx, id)
+		outs, err = vs.history.GetUxOuts(tx, []cipher.SHA256{id})
 		return err
 	}); err != nil {
 		return nil, err
 	}
 
-	return out, nil
+	if len(outs) == 0 {
+		return nil, nil
+	}
+
+	return outs[0], nil
 }
 
 // GetAddrUxOuts gets all the address affected UxOuts.
@@ -1771,11 +1762,72 @@ func (vs *Visor) GetUnspentsOfAddrs(addrs []cipher.Address) (coin.AddressUxOuts,
 	return uxa, nil
 }
 
-// VerifySingleTxnAllConstraints verifies an isolated transaction
-func (vs *Visor) VerifySingleTxnAllConstraints(txn *coin.Transaction) error {
-	return vs.DB.View("VerifySingleTxnAllConstraints", func(tx *dbutil.Tx) error {
-		return vs.Blockchain.VerifySingleTxnAllConstraints(tx, *txn, vs.Config.MaxBlockSize)
-	})
+// VerifyTxnVerbose verifies a transaction, it returns transaction's input uxouts, whether the
+// transaction is confirmed, and error if any
+func (vs *Visor) VerifyTxnVerbose(txn *coin.Transaction) ([]wallet.UxBalance, bool, error) {
+	if err := VerifySingleTxnUserConstraints(*txn); err != nil {
+		return nil, false, err
+	}
+
+	var uxa coin.UxArray
+	var head *coin.SignedBlock
+	var isTxnConfirmed bool
+	if err := vs.DB.View("VerifyTxnVerbose", func(tx *dbutil.Tx) error {
+		var err error
+		head, err = vs.Blockchain.Head(tx)
+		if err != nil {
+			return err
+		}
+
+		uxa, err = vs.Blockchain.Unspent().GetArray(tx, txn.In)
+		switch err.(type) {
+		case nil:
+		case blockdb.ErrUnspentNotExist:
+			uxid := err.(blockdb.ErrUnspentNotExist).UxID
+			// Checks if the transaction is confirmed
+			txnHash := txn.Hash()
+			historyTxn, err := vs.history.GetTransaction(tx, txnHash)
+			if err != nil {
+				return fmt.Errorf("get transaction of %v from historydb failed: %v", txnHash, err)
+			}
+
+			if historyTxn == nil {
+				err = fmt.Errorf("transaction input of %s does not exist in either unspent pool or historydb", uxid)
+				return NewErrTxnViolatesHardConstraint(err)
+			}
+
+			// Transaction is confirmed
+			isTxnConfirmed = true
+			// Gets uxouts of txn.In from historydb
+			outs, err := vs.history.GetUxOuts(tx, txn.In)
+			if err != nil {
+				return err
+			}
+
+			uxa = coin.UxArray{}
+			for _, out := range outs {
+				uxa = append(uxa, out.Out)
+			}
+			return nil
+		default:
+			return err
+		}
+
+		if err := VerifySingleTxnSoftConstraints(*txn, head.Time(), uxa, vs.Config.MaxBlockSize); err != nil {
+			return err
+		}
+
+		return VerifySingleTxnHardConstraints(*txn, head, uxa)
+	}); err != nil {
+		return nil, isTxnConfirmed, err
+	}
+
+	uxs, err := wallet.NewUxBalances(head.Time(), uxa)
+	if err != nil {
+		return nil, isTxnConfirmed, err
+	}
+
+	return uxs, isTxnConfirmed, nil
 }
 
 // AddressCount returns the total number of addresses with unspents
@@ -1844,7 +1896,13 @@ func (vs *Visor) CreateTransactionDeprecated(wltID string, password []byte, coin
 	// Check that the transaction is valid before returning it to the caller.
 	// NOTE: this isn't inside the database transaction, but it's safe,
 	// if a racing database write caused this transaction to be invalid, it would be caught here
-	if err := vs.VerifySingleTxnAllConstraints(txn); err != nil {
+	if err := VerifySingleTxnUserConstraints(*txn); err != nil {
+		logger.WithError(err).Error("Created transaction violates transaction constraints")
+		return nil, err
+	}
+	if err := vs.DB.View("VerifySingleTxnSoftHardConstraints", func(tx *dbutil.Tx) error {
+		return vs.Blockchain.VerifySingleTxnSoftHardConstraints(tx, *txn, vs.Config.MaxBlockSize)
+	}); err != nil {
 		logger.WithError(err).Error("Created transaction violates transaction constraints")
 		return nil, err
 	}
@@ -1901,7 +1959,13 @@ func (vs *Visor) CreateTransaction(params wallet.CreateTransactionParams) (*coin
 	// Check that the transaction is valid before returning it to the caller.
 	// NOTE: this isn't inside the database transaction, but it's safe,
 	// if a racing database write caused this transaction to be invalid, it would be caught here
-	if err := vs.VerifySingleTxnAllConstraints(txn); err != nil {
+	if err := VerifySingleTxnUserConstraints(*txn); err != nil {
+		logger.WithError(err).Error("Created transaction violates transaction constraints")
+		return nil, nil, err
+	}
+	if err := vs.DB.View("VerifySingleTxnSoftHardConstraints", func(tx *dbutil.Tx) error {
+		return vs.Blockchain.VerifySingleTxnSoftHardConstraints(tx, *txn, vs.Config.MaxBlockSize)
+	}); err != nil {
 		logger.WithError(err).Error("Created transaction violates transaction constraints")
 		return nil, nil, err
 	}